Woking on CG method

+++ MatrixAssembleKernel.h
@@ -40,7 +40,7 @@ __device__ void ComputeTangentMatrixDevice(int el_npts, int el_dofs, REAL *stora
     // free(DeBip);
 }
 
-__device__ int64_t me(int64_t *ia_to_sequence, int64_t *ja_to_sequence, int64_t & i_dest, int64_t & j_dest) {
+__device__ int64_t me(int *ia_to_sequence, int *ja_to_sequence, int64_t & i_dest, int64_t & j_dest) {
         
     // Get the matrix entry at (row,col) without bound checking
     int64_t row(i_dest),col(j_dest);
@@ -59,8 +59,8 @@ __device__ int64_t me(int64_t *ia_to_sequence, int64_t *ja_to_sequence, int64_t 
 }
 
 __global__ void MatrixAssembleKernel(int nel, REAL *Kg, int first_el, int64_t *el_color_index, REAL *weight, int *dof_indexes, 
-	REAL *storage, int *rowsizes, int *colsizes, int *rowfirstindex, int *colfirstindex, int *matrixposition, int64_t *ia_to_sequence, int64_t *ja_to_sequence,
-    int64_t *ia_to_sequence_linear, int64_t *ja_to_sequence_linear, REAL *KgLinear) {
+	REAL *storage, int *rowsizes, int *colsizes, int *rowfirstindex, int *colfirstindex, int *matrixposition, int *ia_to_sequence, int *ja_to_sequence,
+    int *ia_to_sequence_linear, int *ja_to_sequence_linear, REAL *KgLinear) {
 
 	int tid = blockIdx.x * blockDim.x + threadIdx.x;
